Limit iterations in getShifts->Get rest of users
This while loop can run infinitely if the number of users in a schedule is less than the total number of users in the org (users.length). 
This is a quick hack, the logic for ending the while loop may need to be fixed, or the logic for obtaining the users.